+++ styles/cart.css
@@ -1,0 +1,4 @@
+table {
+    border: 5px solid rgb(88, 42, 42)
+    
+}
